Update patch set 1
Patch Set 1: Code-Review-1

The line calculate an address, casts it to a pointer, and then wants to access the content of that address. In this patch you set data to the address and not the content of the address ...

Can you please post the error you get from the compiler?

Patch-set: 1
Reviewer: Andreas Schneider <2320@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1